+++ docs/CHANGELOG.md
@@ -24,6 +24,7 @@ General:
    - Added cubzh version 5 (older version) support
    - Added support for writing uncompressed qubicle `qb` files
    - Fixed qubicle `qb` version number
+   - Added support for rooms.xyz `thing` file support
 
 VoxConvert: